feat(pair): ftw-connect friend CLI + wormhole moved to shared internal package
- Extract wormhole subprocess wrapper from go/cmd/ftw-pair/wormhole.go
  into go/internal/wormhole/ (Host, Client, StartHost, Connect) so both
  sidecar and friend CLI can share it without code duplication.
- Replace go/cmd/ftw-pair/wormhole.go with a thin re-export shim that
  preserves the old WormholeHost/WormholeClient/Start*/Connect* names.
- Implement go/cmd/ftw-connect: wormhole tunnel setup, claude mcp add/remove
  lifecycle, clipboard prompt copy, and signal-aware blocking loop.
- Add go/cmd/ftw-connect/clipboard.go (pbcopy / xclip / wl-copy).
- Move wormhole unit tests (splitCompositeCode, pickFreePort, FowlMissing)
  to go/internal/wormhole/wormhole_test.go; ftw-pair test shim tests
  the exported aliases only.
- All 40 packages pass go test ./...
