+++ go/cmd/ftw-connect/main_test.go
@@ -1,17 +1,19 @@
 package main
 
 import (
-	"os/exec"
 	"strings"
 	"testing"
 )
 
-func TestVersionFlag(t *testing.T) {
-	out, err := exec.Command("go", "run", ".", "-version").CombinedOutput()
-	if err != nil {
-		t.Fatalf("run failed: %v\n%s", err, out)
+func TestBuildPrompt(t *testing.T) {
+	p := buildPrompt("write a goodwe driver", "3h45m")
+	if !strings.Contains(p, "goodwe") {
+		t.Fatal("intent missing")
 	}
-	if !strings.Contains(string(out), "ftw-connect") {
-		t.Fatalf("expected ftw-connect, got: %s", out)
+	if !strings.Contains(p, "3h45m") {
+		t.Fatal("ttl missing")
+	}
+	if !strings.Contains(p, "ftw-remote") {
+		t.Fatal("server name missing")
 	}
 }
