+++ go/cmd/ftw-pair/wormhole_test.go
@@ -2,13 +2,13 @@ package main
 
 import (
 	"context"
-	"fmt"
 	"net"
 	"os"
 	"os/exec"
-	"strings"
 	"testing"
 	"time"
+
+	wh "github.com/frahlg/forty-two-watts/go/internal/wormhole"
 )
 
 // TestWormholeForwardEndToEnd performs a real end-to-end forwarding test using
@@ -19,8 +19,8 @@ func TestWormholeForwardEndToEnd(t *testing.T) {
 	if os.Getenv("WORMHOLE_TEST") == "" {
 		t.Skip("set WORMHOLE_TEST=1 to run against real rendezvous (needs internet + fowl)")
 	}
-	if _, err := exec.LookPath(fowldBinary); err != nil {
-		t.Skipf("fowld not installed — `pipx install fowl` to enable this test: %v", err)
+	if _, err := exec.LookPath("fowld"); err != nil {
+		t.Skipf("fowld not installed — `uv tool install fowl` to enable this test: %v", err)
 	}
 
 	// Stand up a local echo server that the wormhole will forward to.
@@ -39,7 +39,7 @@ func TestWormholeForwardEndToEnd(t *testing.T) {
 		c.Write([]byte("PONG\n")) //nolint:errcheck
 	}()
 
-	// Start the host side, forwarding to the echo server.
+	// Start the host side via the shim.
 	ctx := context.Background()
 	host, err := StartWormholeHost(ctx, echoLn.Addr().String())
 	if err != nil {
@@ -48,7 +48,7 @@ func TestWormholeForwardEndToEnd(t *testing.T) {
 	defer host.Close()
 	t.Logf("wormhole code: %s", host.Code)
 
-	// Start the client side using the host's code.
+	// Start the client side via the shim.
 	client, err := ConnectWormholeClient(ctx, host.Code)
 	if err != nil {
 		t.Fatalf("ConnectWormholeClient: %v", err)
@@ -74,18 +74,16 @@ func TestWormholeForwardEndToEnd(t *testing.T) {
 	}
 }
 
-// TestFowlMissing verifies that StartWormholeHost returns ErrFowlNotFound when
-// fowld is not on PATH.  It temporarily overrides PATH to a directory where
-// fowld is unlikely to exist.
+// TestFowlMissing verifies that StartWormholeHost returns *wh.ErrFowlNotFound
+// when fowld is not on PATH.
 func TestFowlMissing(t *testing.T) {
-	// Use a directory that almost certainly does not contain fowld.
 	const emptyPath = "/usr/bin:/bin"
 
 	oldPath := os.Getenv("PATH")
-	os.Setenv("PATH", emptyPath)         //nolint:errcheck
-	defer os.Setenv("PATH", oldPath)     //nolint:errcheck
+	os.Setenv("PATH", emptyPath)     //nolint:errcheck
+	defer os.Setenv("PATH", oldPath) //nolint:errcheck
 
-	if _, err := exec.LookPath(fowldBinary); err == nil {
+	if _, err := exec.LookPath("fowld"); err == nil {
 		t.Skipf("fowld is in %s — cannot test missing-fowl branch", emptyPath)
 	}
 
@@ -93,73 +91,18 @@ func TestFowlMissing(t *testing.T) {
 	if err == nil {
 		t.Fatal("expected error when fowld missing, got nil")
 	}
-	var notFound *ErrFowlNotFound
+	var notFound *wh.ErrFowlNotFound
 	if !isErrFowlNotFound(err, &notFound) {
-		t.Fatalf("expected *ErrFowlNotFound, got %T: %v", err, err)
+		t.Fatalf("expected *wh.ErrFowlNotFound, got %T: %v", err, err)
 	}
 }
 
-// TestSplitCompositeCode validates the code-splitting helper.
-func TestSplitCompositeCode(t *testing.T) {
-	cases := []struct {
-		code        string
-		wantFowl    string
-		wantPort    string
-		wantErrFrag string
-	}{
-		{"7-spinach-atlas:9876", "7-spinach-atlas", "9876", ""},
-		{"2-retrieval-robust:1234", "2-retrieval-robust", "1234", ""},
-		{"nocolon", "", "", "expected"},
-		{":9999", "", "", "non-empty"},
-		{"7-spinach-atlas:", "", "", "non-empty"},
-	}
-	for _, tc := range cases {
-		fowlCode, port, err := splitCompositeCode(tc.code)
-		if tc.wantErrFrag != "" {
-			if err == nil {
-				t.Errorf("splitCompositeCode(%q): expected error containing %q, got nil", tc.code, tc.wantErrFrag)
-				continue
-			}
-			if !strings.Contains(err.Error(), tc.wantErrFrag) {
-				t.Errorf("splitCompositeCode(%q): error %q does not contain %q", tc.code, err, tc.wantErrFrag)
-			}
-			continue
-		}
-		if err != nil {
-			t.Errorf("splitCompositeCode(%q): unexpected error: %v", tc.code, err)
-			continue
-		}
-		if fowlCode != tc.wantFowl || port != tc.wantPort {
-			t.Errorf("splitCompositeCode(%q): got (%q, %q), want (%q, %q)",
-				tc.code, fowlCode, port, tc.wantFowl, tc.wantPort)
-		}
-	}
-}
-
-// TestPickFreePort verifies that pickFreePort returns a usable port.
-func TestPickFreePort(t *testing.T) {
-	port, err := pickFreePort()
-	if err != nil {
-		t.Fatalf("pickFreePort: %v", err)
-	}
-	if port < 1 || port > 65535 {
-		t.Fatalf("pickFreePort returned out-of-range port %d", port)
-	}
-	// Verify we can actually listen on the returned port.
-	ln, err := net.Listen("tcp", fmt.Sprintf("127.0.0.1:%d", port))
-	if err != nil {
-		t.Fatalf("listen on picked port %d: %v", port, err)
-	}
-	ln.Close()
-}
-
-// isErrFowlNotFound is a helper that type-asserts err to *ErrFowlNotFound via
-// errors.As semantics without importing errors (to keep the test file minimal).
-func isErrFowlNotFound(err error, target **ErrFowlNotFound) bool {
+// isErrFowlNotFound type-asserts err to *wh.ErrFowlNotFound.
+func isErrFowlNotFound(err error, target **wh.ErrFowlNotFound) bool {
 	if err == nil {
 		return false
 	}
-	if e, ok := err.(*ErrFowlNotFound); ok {
+	if e, ok := err.(*wh.ErrFowlNotFound); ok {
 		if target != nil {
 			*target = e
 		}
